v2.5.159+235
Functioning deposit and withdraw
Replace double by decimal in many places
Updated number util
Updated localizations
Updated api service

+++ lib/views/wallet/wallet_features/add_gas/add_gas_viewmodel.dart
@@ -111,7 +111,7 @@ class AddGasViewModel extends FutureViewModel {
     setBusy(false);
   }
 
-  checkPass(double amount, context) async {
+  checkPass(Decimal amount, context) async {
     setBusy(true);
     if (isAmountInvalid) {
       sharedService.showInfoFlushbar(
